Added a frontend/GUI
I updated the GUI by implementing the areas where each part of the app will be configured: the teacher, subject and class. This is where their corresponding middle area will be interacting with, I also connected each part to make sure that a change in one will result in a change in the other, it is very dynamic and works well. I used co-pilot to get the styling of the app but a better method will come in future updates. The school and timetable editors will also be ready by the next commit