Massive updates with the timetable editor
I added a new option called the timetable editor where you can edit the position and placement of the individual subjects in the timetable. Apart from that I fixed some other bugs plaguing the code and add some ease of use functionality that makes the app better. For now it compiles the timetable immediately the code is run, but that will change soon.

In future updates I hope to:
1. Implement the theming system to the timetable section
2. Add a generate new timetable and reset timetable button to the timetable option for ease of use and shorter wait times for testing
3. Improve the theme and UI of the app
4. Add the final part of the app which is a designer that allows you to designer where and how each element enters the actual photo of the timetable
5. Connect the classes, teachers and subjects options with the timetable options
6. Add undo and redo functionality
7. Populate the menu bar options with actual stuff
8. Find a way to make the background of the break section be a different colour from the regular