Fixed some crashable problems
I have lost my desire to continue this project, there is always a problem to fix at every point. I want to move on to something else. If I continue with this I will actually start crying. I might come back to this project one day, but for now I need a break for this project. I'm sorry. But I'll be back, one day.