Fixed some of the problems with the teacher and subject
I'm back, I thought about it for about 20 minutes on the toilet and I told myself "You've reached this point, why stop now". So I want to continue working on this project and finish it before the 15th of September when I resume school. Hopefully I would have completed the official Github release of the app by then.