Successfully updated the timetable interaction
I made it so changes made in the settings are directly felt in the timetable and not passively with the refresh button. These are small steps I am making to remove the need for a refresh button. I also fixed some weird save bugs and the "keyboard on-focus on another button will activate that button when any key is clicked" bug